fix : 어드민 파일 분리
