+++ src/basic/components/ui/admin/index.ts
@@ -1,0 +1,5 @@
+export { AdminTabs } from "./AdminTabs";
+export { ProductTable } from "./ProductTable";
+export { ProductForm } from "./ProductForm";
+export { CouponGrid } from "./CouponGrid";
+export { CouponForm } from "./CouponForm";
